requirements.txt
requests>=2.31.0
pandas>=2.0.0
PyJWT>=2.8.0
cryptography>=42.0.0
